Move main function logic to runWarpRouteApply
